requirements.txt
opencv-python==4.11.0.86
python-dotenv==1.0.1
requests==2.32.3